+++ rs-react-app/src/components/search/SearchBlock.tsx
