+++ rs-react-app/src/types/app/appState.ts
@@ -1,4 +1,4 @@
-import type { Card } from './card';
+import type { Card } from '../card';
 
 interface AppState {
   inputValue: string;
